refactors, SP mixin
